+++ app/Filament/Resources/ProductResource.php
@@ -104,10 +104,10 @@ class ProductResource extends BaseResource
 
 namespace App\Filament\Resources\ProductResource\Pages;
 
+use App\Filament\Resources\Pages\CreateRecord;
 use App\Filament\Resources\Pages\EditRecord;
 use App\Filament\Resources\Pages\ListRecords;
 use App\Filament\Resources\ProductResource;
-use Filament\Resources\Pages\CreateRecord;
 
 class ListProducts extends ListRecords
 {